Clean-Up and restructure project
