+++ cmd/main.go
@@ -1,0 +1,25 @@
+package cmd
+
+import (
+	"fmt"
+	"os"
+
+	tea "github.com/charmbracelet/bubbletea"
+	"github.com/zMoooooritz/Nachrichten/pkg/config"
+	"github.com/zMoooooritz/Nachrichten/pkg/tui"
+)
+
+func Run() {
+
+	cfg := config.Init()
+	configuration := cfg.Load()
+
+	p := tea.NewProgram(tui.InitialModel(configuration),
+		tea.WithAltScreen(),
+		// tea.WithMouseCellMotion(),
+	)
+	if _, err := p.Run(); err != nil {
+		fmt.Printf("Alas, there's been an error: %v", err)
+		os.Exit(1)
+	}
+}
